Add very basic subsystem extraction
If subsys_offset is nonzero (or not spaces, yay backwards compatibility!), then
there's an extra unnamed miMATRIX at the end of the file. This matrix contains
the data of a complete matfile (except without the header), containing all the
data for class objects in an undocumented layout.  Crazily, this subsystem can
additionally include another matrix element that contains yet another matfile.