Strip null characters from variable name in `getvarnames` (#210)
This updates `getvarnames` to match the functionality of `read_matrix` using `strip` to deal with the possibility of terminal (or beginning) null characters.

Note: `getvarnames` is only used when 1) checking if a `Matlabv4File` has a specific key, 2) getting the keys of such a file OR 3) `read`ing a specific variable from such a file.